Fix build errors and improve type-checking performance
- Fix SyncController.swift: Convert markdown-style comments to Swift doc comments
- Fix EventMarker.swift: Rename EventType to DrivingEventType to resolve naming conflict
- Fix EventDetector.swift: Update references to use DrivingEventType
- Fix FileScanner.swift: Add explicit UInt64 type conversion for file size
- Fix ContentView.swift: Extract notification handlers to reduce complexity
- Update project.yml: Configure FFmpeg library paths for Homebrew installation

+++ BlackboxPlayer/Services/FileScanner.swift
@@ -629,7 +629,7 @@ class FileScanner {
 
         // 파일 크기 조회
         // FileSystemService.getFileInfo()로 파일 정보 읽기
-        let fileSize = (try? fileSystemService.getFileInfo(at: fileURL).size) ?? 0
+        let fileSize = UInt64((try? fileSystemService.getFileInfo(at: fileURL).size) ?? 0)
 
         // 기본 파일명 생성 (카메라 위치 코드 제외)
         // "20240115_143025_F" → "20240115_143025"
